Added TypeConverter and refactored minor code
